코드 개선
