+++ src/PostProcessor.ts
@@ -1,6 +1,6 @@
 import { MarkdownPostProcessorContext } from "obsidian";
 import LinkThumbnailPlugin from "./main";
-import { linkThumbnailWidgetParams } from "./EnbedDecoratiion";
+import { LinkThumbnailWidgetParams } from "./LinkThumbnailWidgetParams";
 
 export class PostProcessor {
 	plugin: LinkThumbnailPlugin;
@@ -14,17 +14,18 @@ export class PostProcessor {
 		context: MarkdownPostProcessorContext
 	) => {
 		// 링크 변환
-		const linkEls = element.findAll("a.external-link");
+		const linkEls = element.findAll("a.external-link:not(.cm-formatting)");
 		for (const linkEl of linkEls) {
 			// dataview 클래스를 가진 부모 요소를 확인합니다.
 			if (linkEl.closest(".dataview") !== null) {
 				continue;
 			}
-
 			const url = linkEl.innerHTML;
-			const params = await linkThumbnailWidgetParams(url);
+			const params = await LinkThumbnailWidgetParams(url);
 			if (params != null) {
 				linkEl.innerHTML = params;
+				linkEl.removeClass("external-link");
+				linkEl.addClass("markdown-rendered");
 			}
 		}
 	};
